Optionally mount META dir and clean up container
Add an optional command line parameter to specify the path to a META
build folder that should be mounted into the container.

Add '--rm' parameter when running the container to clean up when
exiting. This is safe since all output build artifacts are written to
'/src/out' for external consumption.

Change-Id: Ic2208ea6992e0de5235df923ce45a8f362cc8358