Inject PATH envar into nsjail to eliminate manual "export PATH" command.
By default nsjail does not propagate the environment into the jail. We need the
path to be set up. There are a few ways to solve this problem, but to avoid an
undocumented dependency we are explict about the path we inject. Also add some
comments to old code and fix some documentation to be more consistent with
other areas of the documentation.

Change-Id: I16b6e4585e3ebc3c01b811305f76536effe82b5b